refactor: move property request hooks to a new directory structure
